WIP: Replacing pixel sizes with OpenGL-like scalable units. Font glyphs will be rendered again when framebuffer size changes.

+++ src/xci/text/Text.h
@@ -34,7 +34,7 @@ public:
     void set_string(const std::string& string) { m_string = string; }
     void set_width(float width) { m_layout.set_width(width); }
     void set_font(Font& font) { m_layout.set_font(font); }
-    void set_size(unsigned size) { m_layout.set_size(size); }
+    void set_size(float size) { m_layout.set_size(size); }
     void set_color(const graphics::Color& color) { m_layout.set_color(color); }
 
     void draw(graphics::View& target, const util::Vec2f& pos);
